move npow_mint from operation to staking
