+++ pallets/staking/src/weights.rs
@@ -76,6 +76,7 @@ pub trait WeightInfo {
     fn set_history_depth(e: u32) -> Weight;
     fn reap_stash(s: u32) -> Weight;
     fn new_era(v: u32, d: u32) -> Weight;
+    fn npow_mint() -> Weight;
 }
 
 /// Weights for pallet_staking using the Substrate node and recommended hardware.
@@ -232,6 +233,11 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
             .saturating_add(T::DbWeight::get().reads(15 as Weight))
             .saturating_add(T::DbWeight::get().writes(8 as Weight))
     }
+    fn npow_mint() -> Weight {
+        (33_965_000 as Weight)
+            .saturating_add(T::DbWeight::get().reads(2 as Weight))
+            .saturating_add(T::DbWeight::get().writes(1 as Weight))
+    }
 }
 
 // For backwards compatibility and tests
@@ -386,5 +392,10 @@ impl WeightInfo for () {
         (105_590_000 as Weight)
             .saturating_add(RocksDbWeight::get().reads(15 as Weight))
             .saturating_add(RocksDbWeight::get().writes(8 as Weight))
+    }
+    fn npow_mint() -> Weight {
+        (33_965_000 as Weight)
+            .saturating_add(RocksDbWeight::get().reads(2 as Weight))
+            .saturating_add(RocksDbWeight::get().writes(1 as Weight))
     }
 }
